Move ExceptionListener into the util.error package
