+++ src/main/java/thorpe/luke/network/simulation/worker/Worker.java
@@ -7,8 +7,8 @@ import thorpe.luke.network.simulation.Topology;
 import thorpe.luke.network.simulation.mail.Mailbox;
 import thorpe.luke.network.simulation.mail.PostalService;
 import thorpe.luke.time.Clock;
-import thorpe.luke.util.ExceptionListener;
 import thorpe.luke.util.ThreadNameGenerator;
+import thorpe.luke.util.error.ExceptionListener;
 
 public class Worker {
