+++ src/main/java/thorpe/luke/network/simulation/worker/WorkerManager.java
@@ -14,7 +14,7 @@ import thorpe.luke.network.simulation.Topology;
 import thorpe.luke.network.simulation.mail.Mailbox;
 import thorpe.luke.network.simulation.mail.PostalService;
 import thorpe.luke.time.Clock;
-import thorpe.luke.util.ExceptionListener;
+import thorpe.luke.util.error.ExceptionListener;
 
 public class WorkerManager {
